fix: adds controller to complete MVC, checkDraw and  minMax are still pending.
